ギルドキューの管理システムのエラーハンドリングと管理者コマンド「cleanup」を追加 v2.3.4
・envにOwner項目を追加
・shutdownHandlingを修正
・アクティビティーに現在の起動ステータスを表示するように変更

+++ src/cleanUp.js
@@ -8,8 +8,8 @@ async function cleanupQueue(guildId) {
     if (!serverQueue) return ;
     if (serverQueue.connection && serverQueue.connection.state.status !== "destroyed") serverQueue.connection.destroy();
 
-    clearInterval(serverQueue.time.interval);
-    serverQueue.ffmpegProcess.kill('SIGKILL');
+    if (serverQueue.time.interval) clearInterval(serverQueue.time.interval);
+    if (serverQueue.ffmpegProcess) serverQueue.ffmpegProcess.kill('SIGKILL');
     musicQueue.delete(guildId);
     updateActivity();
     updatePlayingGuild();
